Added custom image annotaion renderer for Chart demo
