+++ CustomSeriesLabels/CustomSeriesLabels/Android/CustomSeriesLabels.Android.csproj
@@ -15,7 +15,7 @@
     <MonoAndroidResourcePrefix>Resources</MonoAndroidResourcePrefix>
     <MonoAndroidAssetsPrefix>Assets</MonoAndroidAssetsPrefix>
     <AndroidUseLatestPlatformSdk>False</AndroidUseLatestPlatformSdk>
-    <TargetFrameworkVersion>v9.0</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v10.0</TargetFrameworkVersion>
     <NuGetPackageImportStamp>
     </NuGetPackageImportStamp>
   </PropertyGroup>
@@ -50,11 +50,17 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Telerik.UI.for.Xamarin" Version="2020.1.318.1" />
-    <PackageReference Include="Xamarin.Forms" Version="4.5.0.495" />
-    <PackageReference Include="SkiaSharp" Version="1.68.1.1" />
-    <PackageReference Include="SkiaSharp.Views " Version="1.68.1.1" />
-    <PackageReference Include="SkiaSharp.Views.Forms " Version="1.68.1.1" />
+    <PackageReference Include="Telerik.UI.for.Xamarin" Version="2020.3.1022.1" />
+    <PackageReference Include="Xamarin.AndroidX.ExifInterface">
+      <Version>1.1.0.5</Version>
+    </PackageReference>
+    <PackageReference Include="Xamarin.AndroidX.MediaRouter">
+      <Version>1.1.0.5</Version>
+    </PackageReference>
+    <PackageReference Include="Xamarin.Forms" Version="4.8.0.1560" />
+    <PackageReference Include="SkiaSharp" Version="2.80.2" />
+    <PackageReference Include="SkiaSharp.Views " Version="2.80.2" />
+    <PackageReference Include="SkiaSharp.Views.Forms " Version="2.80.2" />
     <PackageReference Include="Xamarin.Android.Support.v4" Version="28.0.0.3" />
     <PackageReference Include="Xamarin.Android.Support.CustomTabs" Version="28.0.0.3" />
     <PackageReference Include="Xamarin.Android.Support.v4" Version="28.0.0.1" />
@@ -69,11 +75,13 @@
     <PackageReference Include="Xamarin.Android.Support.v7.RecyclerView" Version="28.0.0.3" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Effects\CustomAnnotationEffect.Android.cs" />
     <Compile Include="Effects\CustomLabelEffect.Android.cs" />
-    <Compile Include="Effects\MyHorizontalLabelRenderer.cs" />
-    <Compile Include="Effects\MyScatterPointRenderer.cs" />
-    <Compile Include="Effects\MyVerticalLabelRenderer.cs" />
+    <Compile Include="ChartFeatureRenderers\MyHorizontalLabelRenderer.cs" />
+    <Compile Include="ChartFeatureRenderers\MyScatterPointRenderer.cs" />
+    <Compile Include="ChartFeatureRenderers\MyVerticalLabelRenderer.cs" />
     <Compile Include="MainActivity.cs" />
+    <Compile Include="ChartFeatureRenderers\ImageAnnotationRenderer.cs" />
     <Compile Include="Resources\Resource.Designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
@@ -110,7 +118,6 @@
       <Name>CustomSeriesLabels.Portable</Name>
     </ProjectReference>
   </ItemGroup>
-  <ItemGroup />
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
      Other similar extension points exist, see Microsoft.Common.targets.
