+++ CustomSeriesLabels/CustomSeriesLabels/Portable/Effects/CustomAnnotationEffect.Forms.cs
@@ -1,0 +1,12 @@
+﻿using Xamarin.Forms;
+
+namespace CustomSeriesLabels.Portable.Effects
+{
+    public class CustomAnnotationEffect : RoutingEffect
+    {
+        public CustomAnnotationEffect()
+            : base("MyCompany.CustomAnnotationEffect")
+        {
+        }
+    }
+}
