+++ CustomSeriesLabels/CustomSeriesLabels/Portable/Views/AnnotationsPage.xaml.cs
@@ -1,0 +1,16 @@
+﻿using CustomSeriesLabels.Portable.ViewModels;
+using Xamarin.Forms;
+using Xamarin.Forms.Xaml;
+
+namespace CustomSeriesLabels.Portable.Views
+{
+    [XamlCompilation(XamlCompilationOptions.Compile)]
+    public partial class AnnotationsPage : ContentPage
+    {
+        public AnnotationsPage()
+        {
+            InitializeComponent();
+            BindingContext = new AnnotationsViewModel();
+        }
+    }
+}
