+++ CustomSeriesLabels/CustomSeriesLabels/iOS/CustomSeriesLabels.iOS.csproj
@@ -120,10 +120,10 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Telerik.UI.for.Xamarin" Version="2020.1.318.1" />
-    <PackageReference Include="Xamarin.Forms" Version="4.5.0.495" />
-    <PackageReference Include="SkiaSharp" Version="1.68.1.1" />
-    <PackageReference Include="SkiaSharp.Views " Version="1.68.1.1" />
-    <PackageReference Include="SkiaSharp.Views.Forms " Version="1.68.1.1" />
+    <PackageReference Include="Xamarin.Forms" Version="4.8.0.1560" />
+    <PackageReference Include="SkiaSharp" Version="2.80.2" />
+    <PackageReference Include="SkiaSharp.Views " Version="2.80.2" />
+    <PackageReference Include="SkiaSharp.Views.Forms " Version="2.80.2" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Portable\CustomSeriesLabels.Portable.csproj">
